Try to get builds on AppVeyor
Recent versions of Visual Studio apparently ship with a snprintf function
and will complain loudly (i.e. #error out) if we define a macro of the same
name.  See http://stackoverflow.com/q/27754492/1468366 for details.
Perhaps it is time to get rid of some of the special cases for Windows?